Moved test to TestViewFsOverloadSchemeListStatus.java
